AppRestApis/requirements.txt
asgiref==3.4.1
Django==3.1.13
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
PyJWT==2.1.0
pymongo==3.12.0
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.2.4
